+++ crates/controller-shared-types/src/macros/mod.rs
@@ -1,3 +1,0 @@
-//! Macros for k3k that are shared among controller and controller modules
-mod redis;
-pub use self::redis::*;