+++ crates/controller/src/lib.rs
@@ -75,7 +75,6 @@ pub mod prelude {
     pub use crate::api::signaling::prelude::*;
     pub use crate::api::Participant;
     pub use crate::redis_wrapper::RedisConnection;
-    pub use crate::{impl_from_redis_value_de, impl_to_redis_args, impl_to_redis_args_se};
 
     // re-export commonly used crates to reduce dependency management in module-crates
     pub use actix_web;
@@ -84,7 +83,6 @@ pub mod prelude {
     pub use aws_sdk_s3;
     pub use bytes;
     pub use chrono;
-    pub use displaydoc;
     pub use futures;
     pub use lapin;
     pub use log;