+++ crates/timer/src/storage/ready_status.rs
@@ -1,28 +1,26 @@
 use anyhow::{Context, Result};
 use controller::prelude::*;
 use controller_shared::ParticipantId;
-use displaydoc::Display;
 use redis::AsyncCommands;
+use redis_args::{FromRedisValue, ToRedisArgs};
 use serde::{Deserialize, Serialize};
 
-#[derive(Display)]
-/// k3k-signaling:room={room_id}:participant::{participant_id}::timer-ready-status
-#[ignore_extra_doc_attributes]
 /// A key to track the participants ready status
+#[derive(ToRedisArgs)]
+#[to_redis_args(
+    fmt = "k3k-signaling:room={room_id}:participant::{participant_id}::timer-ready-status"
+)]
 struct ReadyStatusKey {
     room_id: SignalingRoomId,
     participant_id: ParticipantId,
 }
 
-impl_to_redis_args!(ReadyStatusKey);
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, ToRedisArgs, FromRedisValue)]
+#[to_redis_args(serde)]
+#[from_redis_value(serde)]
 pub struct ReadyStatus {
     pub ready_status: bool,
 }
-
-impl_to_redis_args_se!(&ReadyStatus);
-impl_from_redis_value_de!(ReadyStatus);
 
 /// Set the ready status of a participant
 #[tracing::instrument(name = "meeting_timer_ready_set", skip(redis_conn))]